refactor: move api related code into api folder
