+++ timetree_exporter/api/auth.py
@@ -6,7 +6,7 @@ import uuid
 import logging
 from typing import Union
 import requests
-from timetree_exporter.const import API_BASEURI, API_USER_AGENT
+from timetree_exporter.api.const import API_BASEURI, API_USER_AGENT
 
 logger = logging.getLogger(__name__)
